filter: perform speed computations with doubles
Converting to integer before the sqrt calculation can cause loss of
motion at low speed.

+++ src/filter.c
@@ -139,8 +139,8 @@ tracker_by_offset(struct pointer_accelerator *accel, unsigned int offset)
 static double
 calculate_tracker_velocity(struct pointer_tracker *tracker, uint64_t time)
 {
-	int dx;
-	int dy;
+	double dx;
+	double dy;
 	double distance;
 
 	dx = tracker->dx;
